% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{affineTransform}
\alias{affineTransform}
\title{Perform affine transformation on two-dimensional points.}
\usage{
affineTransform(points, translateX = 0, translateY = 0, rotate = 0, dilate = 1)
}
\arguments{
\item{points}{A N-by-2 data frame representing a set of two-dimensional Cartesian points.}

\item{translateX}{A scalar representing the amount by which to move the points in the X direction.}

\item{rotate}{A scalar representing the angle (in radians) by which to rotate the points.}

\item{dilate}{A scalar representing the amount by which to magnify (or shrink) the point positions relative to the center of mass.}

\item{tranlsateY}{A scalar representing the amount by which to move the points in the Y direction.}
}
\value{

}
\description{
Perform affine transformation on two-dimensional points.
}
