% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calcCorrelationFunction}
\alias{calcCorrelationFunction}
\title{Calculate fluctuation correlations in a full velocity field.}
\usage{
calcCorrelationFunction(velocityField, sampleSize = 2000)
}
\arguments{
\item{velocityField}{A data frame containing all the vectors in a velocity field (full velocities)}

\item{sampleSize}{How many vectors to use from the original field for pair-wise correlation comparisons.}
}
\value{
The correlation functions/profiles for the vector field.
}
\description{
Calculate fluctuation correlations in a full velocity field.
}
