% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calcCorrelationLengths}
\alias{calcCorrelationLengths}
\title{Get correlation lengths from profiles}
\usage{
calcCorrelationLengths(groupedProfiles)
}
\arguments{
\item{correlationProfiles}{A data frame of correlation profiles, grouped by a certain column.}
}
\description{
This function takes a data frame of correlation profiles (grouped by a certain variable) and determines the zero-intercept (correlation length) for each ccorrelation profile.
It is dependent on nesting ahd purrr map functions to iterate over all the profiles. You should install those two packages before using this.
}
