% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calculateFluctuationField}
\alias{calculateFluctuationField}
\title{Calculate fluctuation field from a regular velocity field}
\usage{
calculateFluctuationField(frameVectorField)
}
\arguments{
\item{frameVectorField}{A data frame of 4 columns describing points in space with velocities (2D data)}
}
\value{
The fluctuation field once affine movements are removed.
}
\description{
Calculate fluctuation field from a regular velocity field
}
