% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calculateOptimalAffine}
\alias{calculateOptimalAffine}
\title{Determine affine transformation parameters for best point-wise fit}
\usage{
calculateOptimalAffine(positions)
}
\arguments{
\item{positions}{A N-by-4 data frame containing X1, Y1, X2, Y2, data for 2 sets of 2D Cartesian points.}
}
\value{
Affine x translation, y translation, rotation angle, and dilatation to match points.
}
\description{
Determine affine transformation parameters for best point-wise fit
}
