% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{correlationSlope}
\alias{correlationSlope}
\title{Measure correlation decay by slope at correlation length}
\usage{
correlationSlope(distances, correlations)
}
\arguments{
\item{distances}{A vector representing distance measures}

\item{correlations}{A vector of the corresponding average pair-wise
correlation at that distance.}
}
\value{

}
\description{
This function takes in two vectors that define the correlation profile of any
collective system and measures the slope at the zero-intercept (correlation
length) by using the 5-point stencil method on the five data points closest
to the x-intercept of the correlation profile. Based off of advice from
http://www.theanalysisfactor.com/r-tutorial-4/
}
