% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{getCorrelationLengths}
\alias{getCorrelationLengths}
\title{Title}
\usage{
getCorrelationLengths(frameProfiles, offset = 0)
}
\arguments{
\item{offset}{}
}
\description{
Title
}
