% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{getFirstZeroCrossing}
\alias{getFirstZeroCrossing}
\title{Get first zero crossing of a function}
\usage{
getFirstZeroCrossing(thisVector, distMapping = 1:length(thisVector))
}
\arguments{
\item{thisVector}{A numeric vector}

\item{distMapping}{A numeric vector}
}
\value{

}
\description{
Get first zero crossing of a function
}
