% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{getPairwiseCorrelations}
\alias{getPairwiseCorrelations}
\title{Title}
\usage{
getPairwiseCorrelations(frameData, memoryMaxVectors = 3000)
}
\arguments{
\item{frameData}{a N-by-5 data frame storing N velocity vectors with column names R, C, u, v, and Frame.}

\item{memoryMaxVectors}{the number of vectors to subset the data frame by in order to avoid memory issues.}
}
\value{
The pairwise velocity, direction, and speed correlation for all pairs of vectors found in frameData
}
\description{
Calculates the pair-wise correlations between all vectors in a velocity field.
A vector can be provided to allow within- and between-group comparisons (not implemented).
Vectors can be subsampled to deal with quadratic memory complexity of calculating pair-wise correlations.
}
