% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{measureDeviation}
\alias{measureDeviation}
\title{Measure sum of rms distance between two sets of points after an affine transform.}
\usage{
measureDeviation(transformParams, data)
}
\arguments{
\item{transformParams}{The parameters to use when applying an affine transformation}

\item{data}{A N-by-4 data frame with columns X and Y of first positions and X and Y of comparison positions.}
}
\value{
The sum of pair-wise distances.
}
\description{
Measure sum of rms distance between two sets of points after an affine transform.
}
