% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{normCorrFunction}
\alias{normCorrFunction}
\title{Title}
\usage{
normCorrFunction(domain, correlations, resolution = 1, nknots = 50)
}
\arguments{
\item{frameData}{}
}
\description{
Title
}
