% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{removeRotationalComponent}
\alias{removeRotationalComponent}
\title{Title}
\usage{
removeRotationalComponent(thisFrame)
}
\arguments{
\item{thisFrame}{}
}
\description{
Title
}
