% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{removeTranslationalComponent}
\alias{removeTranslationalComponent}
\title{Title}
\usage{
removeTranslationalComponent(thisFrame)
}
\arguments{
\item{thisFrame}{}
}
\value{

}
\description{
Title
}
