% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{standardizePlacozoaVectors}
\alias{standardizePlacozoaVectors}
\title{Standardize Placozoa data frames.}
\usage{
standardizePlacozoaVectors(dataFrame)
}
\arguments{
\item{dataFrame}{}
}
\description{
Standardize Placozoa data frames.
}
