% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functs.R
\name{test_retrieveCollective}
\alias{test_retrieveCollective}
\title{test_retrieveCollective}
\usage{
test_retrieveCollective(testField)
}
\arguments{
\item{testField}{}
}
\value{

}
\description{
test_retrieveCollective
}
