% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calculateCollectiveField}
\alias{calculateCollectiveField}
\title{Get the collective motion in a vector field.}
\usage{
calculateCollectiveField(frameVectorField)
}
\arguments{
\item{frameVectorField}{A data frame specifying X, Y, vX, and vY for a collectively moving system, where each row is a unit or component.}
}
\value{
A data frame specifying the collective component of the collective movement; i.e. what can be explained by idealized affine transformations.
}
\description{
Get the collective motion in a vector field.
}
