% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{getFirstZeroCrossing}
\alias{getFirstZeroCrossing}
\title{Title}
\usage{
getFirstZeroCrossing(thisVector, distMapping = 1:length(thisVector))
}
\arguments{
\item{distMapping}{}
}
\description{
Title
}
